% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{app_citation}
\alias{app_citation}
\title{Generate a recommended app citation}
\usage{
app_citation(author, year, title, publisher, url,
  heading = "Recommended citation", heading_size = "h3")
}
\arguments{
\item{author}{character string, appears as is.}

\item{year}{year of publication.}

\item{title}{citation title.}

\item{publisher}{the publisher.}

\item{url}{the url of the application.}

\item{heading}{the heading of the citation section, defaults to \code{"Recommended citation"}.}

\item{heading_size}{character, the heading tag, defaults to \code{"h3"}.}
}
\value{
a character string wrapped in \code{shiny::HTML}.
}
\description{
Generate a sub section for an app's information page to provide a recommended citation.
}
\details{
All character strings appear as is. For example, enter \code{author="First Last, Last, F. , Last, F"} if that is how
three authors should appear in the citation. Note that terminal periods (\code{.}) are implicit.
}
\examples{
#not run
}
