% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{brushed_data}
\alias{brushed_data}
\title{Filter data frame based on plot brush}
\usage{
brushed_data(data, variable, x, brush)
}
\arguments{
\item{data}{data frame.}

\item{variable, }{character, name of x-axis variable in \code{data}. May be numeric or categorical.}

\item{x}{x coordinates. See details.}

\item{brush}{brushed coordinates. See details.}
}
\value{
a data frame.
}
\description{
This function filters data frame rows to a subset corresponding to observations included in the brushed region of a plot.
}
\details{
This implementation is for brushing restricted to the x-axis.
#' \code{x} and \code{brush} are commonly reactive values in \code{server.R}, storing the observed click and brush coordinates in \code{input},
rather than the \code{input} objects directly.
}
\examples{
#not run
}
