% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{color_indexer}
\alias{color_indexer}
\title{Color re-indexing}
\usage{
color_indexer(x, clrby)
}
\arguments{
\item{x}{a data frame.}

\item{clrby}{column name of the grouping variable to be used for coloring.}
}
\value{
a re-indexed vector of colors from \code{tolpal}.
}
\description{
Re-index colors for grouped data, as displayed in plots and other widgets, when groups may be missing.
}
\details{
This is typically used when group levels are in a data set but user selection (mouse brush action) in an interactive plot may not result
in a data subset that includes data from all levels of the grouping variable.
}
\examples{
# not run
}
