% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{colorFacet}
\alias{colorFacet}
\title{Coloring and faceting helper for ggplot}
\usage{
colorFacet(g, d, cby, clr, fby, scales)
}
\arguments{
\item{g}{a ggplot object.}

\item{d}{a data frame.}

\item{cby}{character, factor column in \code{d}.}

\item{clr}{character, color vector. If \code{NULL}, no coloring by levels of \code{cby} occurs.}

\item{fby}{character, factor column in \code{d}.}

\item{scales}{character, passed to \code{scales} in \code{ggplot2::facet_wrap}.}
}
\value{
a ggplot object.
}
\description{
This function takes a \code{ggplot} object and source data frame, \code{d}, and adds manual color and fill scale layers
based on \code{cby} and \code{clr} and/or faceting based on \code{fby} and \code{scales}.
The data frame, \code{d}, is passed for determining the levels of the factor \code{cby} in \code{d}.
The \code{cby} and \code{fby} columns in \code{d} must be factors.
Like coloring, faceting must be done by a single factor variable, not crossed factors.
}
\examples{
#not run
}
