% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{dashboard_footer}
\alias{dashboard_footer}
\title{Add footer to Shiny dashboard sidebar}
\usage{
dashboard_footer(href, src, label = "", width = "100\%",
  height = "160px", italic = TRUE, bold = TRUE,
  style = "text-align:center; align: center; padding: 0px; margin: 0px;")
}
\arguments{
\item{href}{url, will open in new window.}

\item{src}{image url.}

\item{label}{character, should be short.}

\item{width}{width of logo.}

\item{height}{height of footer, defaults to \code{"160px"}.}

\item{italic}{logical, for label text.}

\item{bold}{logical, for label text.}

\item{style}{character, option style string for logo and label. See function for defaults.}
}
\value{
an html footer tag.
}
\description{
Add a foot to a Shiny dashboard sidebar that includes a logo image link and a label.
}
\details{
\code{style} pertains to the logo and label whereas \code{width} is for the logo specifically and only \code{height}
pertains directly to the footer tag.
}
\examples{
#not run
}
