% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apputils.R
\name{exApp}
\alias{exApp}
\title{Run app examples}
\usage{
exApp(id, source = "local", local_mode = "normal")
}
\arguments{
\item{id}{character, the name of the application. See details.}

\item{source}{character, source of app. See details.}

\item{local_mode}{character, display mode.}
}
\description{
Launch an apputils Shiny app example in your browser.
}
\details{
The \code{source} defaults to \code{"local"} for package apps. If set to \code{"remote"},
\code{snapp} will launch the app in your browser using the canonical url rather than the package app.
Options for \code{local_mode} are \code{"normal"} (default) or \code{"showcase"}
for tandem reactive code highlighting, and metadata if applicable. \code{local_mode} applies to \code{source = "local"} package apps.

Available IDs include \code{"icons"}. This app displays all the available statistics and data analysis themed icons
available in apputils in both light and dark color styles.
}
\examples{
\dontrun{exApp("icons")}
}
