% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{faq}
\alias{faq}
\title{Generate a Frequently Asked Questions (FAQ) widget}
\usage{
faq(faqlist, id, format = "bscollapse", bscollapse_args = list(id = "faq",
  open = NULL, multiple = FALSE))
}
\arguments{
\item{faqlist}{a prepared list of FAQs. See details.}

\item{id}{character, the id for the \code{shinyBS::bsCollapse} element.}

\item{format}{character, \code{"bscollapse"} (default) or alternatively, \code{"list"}.}

\item{bscollapse_args}{a list of arguments passed to bsCollapse. Defaults to \code{list(id = "faq", open = NULL, multiple = FALSE)}.}
}
\value{
a \code{shinyBS::bsCollapse} object or a list.
}
\description{
Generate a widget that displays FAQs in an information/about section of an app.
}
\details{
\code{faq} creates an FAQ widget from a named list.
It returns a Bootstrap Collapse menu FAQ format by default. Changing to \code{format = "list"}
simply passes through the original \code{faqlist}.

Generally, an FAQ list is a named list that is prepared in advance and passed to \code{faqlist}.
The names of the list elements are used for indexing the list with \code{id}.
It is designed this way in the context of reusable FAQ list dictionaries, where each use does not require all entries.
Each list element in \code{faqlist} is a length-2 list.
The first element is the question: a character string. The second element
is the answer: any text or HTML content, e.g., multiple tags wrapped in a tag list.

A key difference with the \code{open} argument that can be passed in a list to \code{bscollapse_args}
is that it must simply match a single \emph{dictionary entry} \code{id} (e.g., \code{open = "apps"}),
not a collapse panel question title or id.
}
\examples{
#not run
}
