% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_breaks}
\alias{get_breaks}
\title{Create breaks for ggplot}
\usage{
get_breaks(yrs, n.facets = 1)
}
\arguments{
\item{yrs}{years vector for time series plot.}

\item{n.facets}{number of facets.}
}
\value{
a numeric vector.
}
\description{
Create breaks for ggplot
}
\examples{
#not run
}
