% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getPosition}
\alias{getPosition}
\title{Positioning helper for ggplot}
\usage{
getPosition(jitter, cby, w = 0.2, h = 0, wd = 0.75, dodgeable = FALSE)
}
\arguments{
\item{jitter}{logical.}

\item{cby}{character, factor to color by.}

\item{w}{width.}

\item{h}{height.}

\item{wd}{dodge width.}

\item{dodgeable}{logical.}
}
\value{
a ggplot position function.
}
\description{
This functions assists with position by combining information on whether or not jitter is present in a plot,
there is a categorical variable for coloring, and if the plot uses dodge positioning when coloring variable levels.
}
\examples{
#not run
}
