% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{gist_url}
\alias{gist_url}
\title{Grab gist urls.}
\usage{
gist_url(id)
}
\arguments{
\item{id}{shorthand character id refering to entry in gist_url dictionary.}
}
\value{
a character string.
}
\description{
Grab gist urls.
}
\examples{
#not run
}
