% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{icon}
\alias{icon}
\title{Override shiny icon function}
\usage{
icon(name, class = NULL, lib = "font-awesome")
}
\arguments{
\item{name}{character. If \code{lib="local"}, \code{name} must be a named list with \code{src} element.}

\item{class}{character.}

\item{lib}{library, may be 'local'.}
}
\value{
an icon tag.
}
\description{
This and the apputils::valueBox functions override functions in the shiny and shinydashboard packages
to provide the ability to display image file icons in value boxes.
}
\examples{
#not run
}
