% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{infoBox}
\alias{infoBox}
\title{Override shinydashboard infoBox function}
\usage{
infoBox(title, value = NULL, subtitle = NULL, icon = NULL,
  color = "aqua", width = 4, href = NULL, fill = FALSE,
  validate.color = FALSE)
}
\arguments{
\item{title}{title.}

\item{value}{value.}

\item{subtitle}{subtitle.}

\item{icon}{icon may be a local image file url.}

\item{color}{character.}

\item{width}{column width.}

\item{href}{link.}

\item{fill}{logical, full color info box.}

\item{validate.color}{bypass \code{shinydashboard:::validateColor}.}
}
\value{
an infoBox.
}
\description{
This and the apputils::icon functions override functions in the shiny and shinydashboard packages
to provide the ability to display image file icons in info boxes.
}
\examples{
#not run
}
