% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apputils.R
\name{inputs_not_null}
\alias{inputs_not_null}
\title{Require non-null inputs in app UI}
\usage{
inputs_not_null(inputs)
}
\arguments{
\item{inputs}{a character vector of inputs.}
}
\value{
a character string of the javascript condition.
}
\description{
Generate quoted javascript logic for non-null inputs.
}
\examples{
inputs_not_null(c("input1", "input2"))
}
