% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{kilo_mega}
\alias{kilo_mega}
\title{Kilo and Mega labels for rescaled data.}
\usage{
kilo_mega(x)
}
\arguments{
\item{x}{numeric value.}
}
\value{
\code{x} as a character string, potentially in the form of a smaller value with a \code{K} or \code{M} suffix.
}
\description{
Kilo (K) and Mega (M) labels for rescaled plot axes and stat boxes.
}
\details{
When data are rescaled, labels in plot axes and other places can be replaced with a shorthand form.
}
\examples{
#not run
}
