% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbox.R
\name{makeIcons}
\alias{makeIcons}
\title{Generate png thumbnail icons for various quantities}
\usage{
makeIcons(primary_color = "#FFFFFF", secondary_color = "#FFFFFF75",
  color_theme = "white", dir = ".")
}
\arguments{
\item{primary_color}{primary color.}

\item{secondary_color}{secondary color.}

\item{color_theme}{color theme.}

\item{dir}{output directory.}
}
\value{
This function is used for its png file-generating side effects.
}
\description{
Generate png thumbnail icons for various quantities including common statistics, model parameter estimates,
trends and rates of change.
}
\details{
The standard white and dark theme icon versions are produced using the default call, \code{makeIcons()},
and \code{makeIcons("black", "gray30", "black")}, respectively.
The function code is repetitive and includes hardcoded values because each icon is uniquely tailored,
for example, for proper text placement. This function serves as a package reference to how the standard icon set
is created from scratch.
}
\examples{
#not run
}
