% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mouseLog}
\alias{mouseLog}
\title{Mouseover info browser feedback}
\usage{
mouseLog(x, ns, width)
}
\arguments{
\item{x}{logical.}

\item{ns}{module namespace.}

\item{width}{Shiny column width.}
}
\value{
a shiny column containing mouseover plot interaction coordinate feedback.
}
\description{
Mouseover info browser feedback
}
\examples{
#not run
}
