% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{pasteByLength}
\alias{pasteByLength}
\title{Create sensible readable string from a vector}
\usage{
pasteByLength(x)
}
\arguments{
\item{x}{character.}
}
\value{
character.
}
\description{
Create a string from a vector for printing where vector elements
are sensibly pasted together, collapsing using commas and \code{'and'} to conjoin vector elements.
}
\examples{
pasteByLength(c("a", "b", "c"))
}
