% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{pTextSize}
\alias{pTextSize}
\title{Size of text in overridden shiny dashboard value box function widget output}
\usage{
pTextSize(x, value, margin = NULL, default.value = 100)
}
\arguments{
\item{x}{text.}

\item{value}{size.}

\item{margin}{numeric, html margin size in pixels.}

\item{default.value}{value to repeat if necessary.}
}
\value{
a shiny taglist.
}
\description{
Size of text in overridden shiny dashboard value box function widget output
}
\examples{
#not run
}
