% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{read_md_paragraphs}
\alias{read_md_paragraphs}
\title{Read paragraphs from file}
\usage{
read_md_paragraphs(file, ptag = FALSE, collapse = FALSE)
}
\arguments{
\item{file}{character, file to read text from.}

\item{ptag}{logical, wrap paragraphs in html paragraph tag an include full justifcation.}

\item{collapse}{logical, collapse vector of paragraphs into a single element.}
}
\value{
a vector of character strings.
}
\description{
Read paragrpahs from a text file based on markdown format.
}
\details{
The text file does not need to be a \code{.md} file. The function merely relies on the assumption that
paragraphs are separated by an empty line, based on markdown formatting. This allows for typing up
paragraphs in a plain text file using this convenient formatting, to be read into R wherever paragraphs of
text are needed in an app.

Common uses include returning a vector of strings to be used as text for each tour step description
in an introjs tour (\code{ptag=FALSE} and \code{collapse=FALSE}, the default)
or a collapsed string of html text (\code{ptag=TRUE} and \code{collapse=TRUE}) for an app description.
}
\examples{
#not run
}
