% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{samplesize_factor}
\alias{samplesize_factor}
\title{Divide sample size by a factor based on the number of RCP-GCM combinations}
\usage{
samplesize_factor(x, baseline)
}
\arguments{
\item{x}{special data frame.}

\item{baseline}{character, name of baseline data set.}
}
\value{
a number.
}
\description{
This function reduces sample size of point sample overlays in plots to improve aesthetics and/or save time
when there are too many points to plot.
}
\examples{
#not run
}
