% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{sbg_to_tables}
\alias{sbg_to_tables}
\title{Convert stat box groups output to list of data frames}
\usage{
sbg_to_tables(d, grp, rnd = 0)
}
\arguments{
\item{d}{data frame to use for generating stat box groups.}

\item{grp}{character or \code{NULL}, factor variable (column name) in \code{d}
used to determine the number of groups of stat box sets.}

\item{rnd}{numeric, numer of decimal places to round values to.}
}
\value{
a list of data frames.
}
\description{
Convert stat box groups output to list of data frames for reuse in other contexts,
typically in reports generated by \code{rmarkdown} and \code{knitr}.
}
\examples{
#not run
}
