% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statboxes.R
\name{stat_boxes}
\alias{stat_boxes}
\title{Generate a set of stat boxes.}
\usage{
stat_boxes(x, type = "annual", style = "valueBox", rnd = 0,
  clrs = c("light-blue", "blue"), dec,
  main_title = "<h4>Aggregate period statistics</h4>", ...)
}
\arguments{
\item{x}{a data frame.}

\item{type}{character, type of stat box set. Must be \code{"annual"} (default) or \code{"decadal"}.}

\item{style}{character, \code{"valueBox"} (default) or \code{"infoBox"}.}

\item{rnd}{decimal places to round data to for appearance in stat box (\code{valueBox} widget). Defaults to \code{0}.}

\item{clrs}{character, background colors of stat boxes. See details.}

\item{dec}{character, the sorted unique values of \code{x$Decade}. Typically provided by calling function. See \code{stat_boxes_group}.}

\item{main_title}{character, typically contains html. Defaults to \code{"<h4>Aggregate period statistics</h4>"}.}

\item{...}{additional arguments passed to valueBox or infoBox. See details.}
}
\value{
a \code{shiny::tagList} containing a heading and a fluid row of six stat boxes.
}
\description{
Generate a set of six related stat boxes that summarize the distribution of values from a data frame.
}
\details{
\code{clrs} must be length-one (will be repeated six times), -two (alternating colors), or -six (all colors individually specified).
Colors must be provided in hex format, e.g., \code{#FFFFFF} or a valid Shiny Dashboard color.
This function is typically called by \code{stat_boxes_group} but may be called directly when groups of stat boxes are not needed.
\code{...} arguments may include \code{src}, a character vector of paths to six local image files to be used as valueBox icons.
Standard icons for annual and decadal stat box sets are used by default. \code{w} for icon width, if not provided, defaults to \code{"90px"}.
\code{height} applies to the overall box height in pixels (\code{"110px"} default), not the icon.
}
\examples{
#not run
}
