% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statboxes.R
\name{stat_boxes_group}
\alias{stat_boxes_group}
\title{Generate a group of stat boxes sets.}
\usage{
stat_boxes_group(x, clrby, clrs.default = list(c("light-blue", "blue")),
  type = "annual", rnd = 0,
  main_title = "<h4>Aggregate period statistics</h4>", prevent, ...)
}
\arguments{
\item{x}{a  master data frame, which may or may not be split into multiple groups, depending on \code{clrby}.}

\item{clrby}{character, column name of \code{x} to color groups of stat box sets by. May be \code{NULL}.}

\item{clrs.default}{list of color vectors, one vector for each group.}

\item{type}{character, type of stat box set. Must be \code{"annual"} (default) or \code{"decadal"}.}

\item{rnd}{decimal places to round data to for appearance in stat box (\code{valueBox} widget). Defaults to \code{0}.}

\item{main_title}{character, typically contains html. Defaults to \code{"<h4>Aggregate period statistics</h4>"}.}

\item{prevent}{logical, whether stat boxes should be prevented (contextual, e.g., no data available in app).}

\item{...}{additional arguments passed to \code{stat_boxes}. See details.}
}
\value{
a group list of tag lists containing stat box sets.
}
\description{
Generate a group of multiple sets of six related stat boxes that summarize the distribution of values from a multiple data frames.
}
\details{
\code{...} arguments may include \code{src}, a character vector of paths to six local image files to be used as valueBox icons.
Standard icons for annual and decadal stat box sets are used by default. \code{w} for icon width, if not provided, defaults to \code{"90px"}.
}
\examples{
#not run
}
