% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statboxes.R
\name{statIcon}
\alias{statIcon}
\title{Accessor for apputils icons}
\usage{
statIcon(id, theme = "white")
}
\arguments{
\item{id}{character, id to be matched to a stat box icon.}

\item{theme}{character, \code{"white"} (default) or \code{"black"}.}
}
\value{
character string specifying the local image file location in the \code{apputils} package.
}
\description{
Access path to local \code{apputils} stat box icon image files.
}
\details{
This accessor function provides convenient access to local package resources.
Specifically, given an ID, it returns the resource path to a stat box icon png file.
This is typically used in a Shiny app that has \code{apputils} loaded.
Available options include:
\describe{
  \item{normal}{Normal distribution density curve}
  \item{min}{Normal curve with minimum line and statistic notation}
  \item{max}{As above, for the maximum}
  \item{mean}{As above, for the mean}
  \item{sd}{As above, with two lines marking +/- one standard deviation}
  \item{median}{As above, for the median}
  \item{iqr}{Box plot with interquartile range marked and labeled}
  \item{b0}{Beta zero hat for linear models, scatter plot with linear model intercept highlight}
  \item{b1}{Beta one hat for linear models, scatter plot with linear model slope highlight}
  \item{r2}{R-squared, scatter plot with correlation highlight}
  \item{pvalue}{Normal distribution curve with lower tail highlight}
  \item{inc}{Increasing trend line with delta change symbol}
  \item{dec}{As above but decreasing}
  \item{pctinc}{As above but with percentage symbol added}
  \item{pctdec}{As above but decreasing and with percentage symbol added}
  \item{barinc}{Increasing consecutive bars with arrow change indicators}
  \item{bardec}{As above but decreasing}
}
}
\examples{
statIcon("normal")
}
