% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableRowColors}
\alias{tableRowColors}
\title{COlor table rows according to ggplot coloring by groups.}
\usage{
tableRowColors(data, variable, colorvec, alpha_append = NULL)
}
\arguments{
\item{data}{a special data frame.}

\item{variable}{a column name in the data specifying the factor used for plot colors.}

\item{colorvec}{a vector of colors.}

\item{alpha_append}{NULL or a two digit numeric or character to append to hex colors.}
}
\value{
a DT::styleEqual call.
}
\description{
COlor table rows according to ggplot coloring by groups.
}
\examples{
#not run
}
