% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{tour_changeClass}
\alias{tour_changeClass}
\title{Generate a JS callback string for rintrojs tour}
\usage{
tour_changeClass(step, from, to)
}
\arguments{
\item{step}{numeric, a vector of steps to apply the change to.}

\item{from}{character, a vector of the the active classes to be removed.}

\item{to}{character, a single class to add as active and trigger with a click.}
}
\value{
a character string.
}
\description{
This function generates an JS callback character string to change remove an active class
and add and trigger a new class in its place.
}
\details{
This function is useful on rintrojs tours in order to, for example, ensure that the active tab panel
changes accordingly depending on tour content that is required to be visible during particular steps.
Note that only one active class should be added at a time. Call twice to add two active classes for a common step.
}
\examples{
#not run
}
