% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apputils.R
\name{update_toastr_css}
\alias{update_toastr_css}
\title{Update shinytoastr css}
\usage{
update_toastr_css(container = NULL, toast = NULL, rgba = NULL,
  hover.rgba = NULL, opacity = NULL, hover.opacity = NULL,
  position = "top-center")
}
\arguments{
\item{container}{list of style arguments for the container div. See details and example.}

\item{toast}{list of style arguments for the toast. See details and example.}

\item{rgba}{numeric, vector of four css rgba property values for background color, e.g., \code{c(0,0,0,0.5)}. See details.}

\item{hover.rgba}{numeric, vector of four css rgba property values for background color on mouse hover. See details.}

\item{opacity}{numeric, toast opacity. Appended to \code{container}.}

\item{hover.opacity}{numeric, toast opacity on mouse hover.}

\item{position}{character, defaults to \code{"top-center"}.}
}
\value{
an html style tag.
}
\description{
Update toast css from shinytoastr package.
}
\details{
\code{apputils} already contains some toastr css overrides (loaded via \code{use_apputils}).
This function allows for injecting additonal or different css overrides for a specific toast container
that may not already be as specified by \code{apputils}. This is typically used to adjust the app intro toast,
hence the default for \code{position} is \code{"top-center"}.

Note that list names and values ay be quoted if necessary. See example.
Should be familiar with source toastr css in addition to running the example
in order to understand which elements apply to \code{container} vs. \code{toast}.

If wanting to keep text fully opaque in the toast while using semi-transparency,
especially useful when adding a background image, use css rgba instead of opacity.
\code{rgba} and\code{hover.rgba} nullify opacity arguments if both are provided, respectively.
}
\examples{
update_toastr_css(
  list('overflow-y'='auto', width='70\%', height='700px'),
  list(top='100px', margin='0 auto', left='115px')
)
}
