% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{valueBox}
\alias{valueBox}
\title{Override shinydashboard valueBox function}
\usage{
valueBox(value, subtitle, icon = NULL, color = "aqua", width = 4,
  href = NULL)
}
\arguments{
\item{value}{value.}

\item{subtitle}{subtitle.}

\item{icon}{icon may be a local image file url.}

\item{color}{character.}

\item{width}{column width.}

\item{href}{link.}
}
\value{
a valueBox.
}
\description{
This and the apputils::icon functions override functions in the shiny and shinydashboard packages
to provide the ability to display image file icons in value boxes.
}
\examples{
#not run
}
