% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{valueBoxColorsCSS}
\alias{valueBoxColorsCSS}
\title{Genrate CSS for different colored value boxes.}
\usage{
valueBoxColorsCSS(bg = "#333333", col = "#FFFFFF")
}
\arguments{
\item{bg}{character, hex color of value box background.}

\item{col}{character, hex color of value box text.}
}
\value{
a character string wrapped in \code{shiny::HTML}.
}
\description{
Generate CSS for background color and text color of value boxes based on hex color input.
}
\details{
Colors must be in hex color format, e.g., \code{"#FFFFFF"}.
This function generates CSS to be added to a Shiny app.
New value box elements are named by appending the hex color to \code{.bg-} without the \code{#}.
When using the \code{valueBox} override function in \code{apputils},
make sure to specify a hex color "name" that is a character string without the \code{#}, e.g., "FFFFFF".
The only available colors that will work with \code{valueBox} in an app are those that have their CSS generated and loaded in the app as well.
}
\examples{
#not run
}
