% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{valueBoxPalette}
\alias{valueBoxPalette}
\title{Generate CSS for a value box color palette.}
\usage{
valueBoxPalette(pal)
}
\arguments{
\item{pal}{character, a named color palette. See details for options.}
}
\value{
a CSS snippet defining value boxes of different colors.
}
\description{
Obtain a CSS snippet defining value box colors for a full color palette from a named list of available palettes.
}
\details{
The only available palette at this time is \code{"tolpal"}, which covers all colors available using the \code{tolpal} function.
See \code{valueBoxColorsCSS} for more information.
}
\examples{
#not run
}
