% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apputils.R
\name{app_overlay}
\alias{app_overlay}
\title{Title}
\usage{
app_overlay(text, logo = NULL, loading.logo = NULL,
  text.loading = "<h1>Loading...</h1>")
}
\arguments{
\item{text}{character, typically a string of html.}

\item{logo}{character, path to image.}

\item{loading.logo}{character, path to image.}

\item{text.loading}{character, text to accompany \code{loading.logo} or alone.}
}
\value{
a \code{shiny::HTML} string.
}
\description{
Title
}
\examples{
#not run
}
