% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{app_showcase}
\alias{app_showcase}
\title{Genrate app showcase content}
\usage{
app_showcase(drop = NULL)
}
\arguments{
\item{drop}{character vector of any apps to exclude from showcase.}
}
\value{
a shiny fluidRow containing organized and stylized app image links for reference.
}
\description{
Genrate app showcase content
}
\examples{
#not run
}
