% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{build_mapset}
\alias{build_mapset}
\title{Build leaflet map from among multiple available maps}
\usage{
build_mapset(shp, all_regions, ids, mapset, mapsets, default = "",
  default.name = default, all_locs_shapefile, xyz, color = c("#000000",
  "#000000", "#3C8DBC"), fill = c("#000000", "#3C8DBC", "#3C8DBC"),
  color_opacity = c(0.5, 1, 1), fill_opacity = c(0, 0, 0.4), progress = F)
}
\arguments{
\item{shp}{a shapefile pertaining to a single \code{mapset}.}

\item{all_regions}{a named vector of region IDs pertaining to regions in \code{shp}.}

\item{ids}{literal names of regions in \code{shp} (may not match \code{regions} elements).}

\item{mapset}{character, name of a mapset in \code{mapsets}.}

\item{mapsets}{character, vector of all available map sets.}

\item{default}{name of a default map set, if used. Otherwise defaults to \code{default=""}.}

\item{default.name}{label used in place of \code{default} string. Defaults to \code{default}.}

\item{all_locs_shapefile}{a named list (names match \code{mapsets})
where each list element is a vector of regions specified to match the region IDs in \code{shp}.}

\item{xyz}{a named list (names match \code{mapsets})
where each list element is a vector of three values: longitude, latitude, and zoom level, for centering and zoom level setup.
Alternatively, a non-list vector of three values if all map sets will share a common centering and zoom level.}

\item{color}{vector of three polygon border colors. See details.}

\item{fill}{vector of three polygon fill colors. See details.}

\item{color_opacity}{numeric vector of three polygon border color opacities. See details.}

\item{fill_opacity}{numeric vector of three polygon fill color opacities. See details.}

\item{progress}{logical, add progress bar in app during map rendering. Helpful is rendering is slow.}
}
\value{
a leaflet map
}
\description{
Build a leaflet map from a shapefile in a context of multiple map sets and corresponding shapefiles.
}
\details{
This function builds a leaflet map from a shapefile. There can be a list of shapefiles, and corresponding
lists of map set names and parameters for setting up each map such as centering and zooming defaults.

Each \code{mapset} and \code{shp} may contain named subregions/polygons. Polygons have fill color, outline, and
change appearance on mouse hover and click (selection).

A \code{default} map may be identified, and if so, it is assumed to be a simpler base map with only a single polygon.
Such a map set has one region, and therefore is fully colored as permanently "selected", with no need to click a polygon.

Polygon border and fill colors and opacities are vectors of three values each.
These correspond to unselected, highlighted and selected polygons, respectively.
While \code{build_mapset} deals primarily with unselected or highlighted polygons and \code{update_mapset}
deals strictly with selected polygons, \code{build_mapset} still takes a third value because this is used
in the case of a \code{default} map set consisting of a single polygon (which is always colored and filled as a "selected" polygon).
}
\examples{
#not run
}
