% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{contactinfo}
\alias{contactinfo}
\title{Add contact information widget to app.}
\usage{
contactinfo(logos = NULL)
}
\arguments{
\item{logos}{named list. Currently accepted names must be \code{uaf} or \code{snap}. List elements are image urls for \code{src}.}
}
\value{
a shiny taglist.
}
\description{
Add contact information widget to app.
}
\examples{
#not run
}
