% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggObserve}
\alias{ggObserve}
\title{Plot Observers}
\usage{
ggObserve(session, dblclick, brush, dbrush, rvx, rvy = NULL,
  rvbrush = brush, zoomable = TRUE)
}
\arguments{
\item{session}{Shiny session object.}

\item{dblclick}{charatcer, double click input id specified in `plotOutput`.}

\item{brush}{charatcer, brush input id specified in `plotOutput`.}

\item{dbrush}{reactive object containing data frame whose rows are based on the brushed region of the plot. See details.}

\item{rvx}{charatcer, name of x coordinates object stored in a reactive values object. See details.}

\item{rvy}{charatcer, name of y coordinates object stored in a reactive values object. See details.}

\item{rvbrush}{charatcer, double click input id specified in `plotOutput`.}

\item{zoomable}{logical, whether the plot is intended to be zoomable on double click. Defaults to \code{TRUE}.}
}
\value{
this function is called for its observer side effects.
}
\description{
Convenient wrapper around observers related to Shiny plot outputs.
}
\details{
These observers require the app plots to be ggplot2 plots.
Note that `resetOnNew` in `plotOutput` should be set to \code{FALSE}.
It may be preferable to leave \code{resetOnNew=TRUE} if setting \code{zoomable=FALSE} here.
These observers require a context where it can also be assumed that plot outputs are
properly isolated and responsive to updates initiated by the observers.
The plot functions themselves also take two data arguments, a brushed data frame that is a subset of rows from
a master data frame, and the master. Internally, these plots functions make their own determination
of how to plot brushed vs. unbrushed data.

The reactive object passed via \code{dbrush} must be wrapped in \code{isolate}.
The reactive values object containing \code{rvx} and \code{rvy} must be named \code{rv_plots},
i.e., \code{rv_plots <- reactiveValues(x=NULL, y=NULL)}.
The reactive values object containing \code{rvbrush} must be named \code{rv},
i.e., \code{rv <- reactiveValues(brush=NULL)}.
\code{zoomable} is best set to \code{FALSE} for discrete x-axis plots like \code{geom_boxplot}
because zooming in on the discrete axis is problematic for the current brushed data observation implementation.
}
\examples{
#not run
}
