% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{interact}
\alias{interact}
\title{Interaction in ggplot.}
\usage{
interact(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
an unevaluated string for ggplot interaction term.
}
\description{
Interaction in ggplot.
}
\examples{
#not run
}
