% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mouseInfo}
\alias{mouseInfo}
\title{Mouseover info}
\usage{
mouseInfo(clk, dblclk, hov, brush)
}
\arguments{
\item{clk}{click coordinates.}

\item{dblclk}{double click coordinates.}

\item{hov}{hover coordinates.}

\item{brush}{brush coordinates.}
}
\value{
character string.
}
\description{
Mouseover info
}
\examples{
#not run
}
