% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statboxes.R
\name{stat_boxes}
\alias{stat_boxes}
\title{Generate a set of stat boxes.}
\usage{
stat_boxes(x, type = "annual", style = "valueBox", rnd = 0, dec,
  height = "110px", width.icon = "90px", text.size = 75,
  value.size = 150, output = "boxes",
  main_title = "<h4>Aggregate period statistics</h4>",
  clrs = c("light-blue", "blue"))
}
\arguments{
\item{x}{a data frame.}

\item{type}{character, type of stat box set. Must be \code{"annual"} (default) or \code{"decadal"}.}

\item{style}{character, \code{"valueBox"} (default) or \code{"infoBox"}.}

\item{rnd}{decimal places to round data to for appearance in stat box (\code{valueBox} widget). Defaults to \code{0}.}

\item{dec}{character, the sorted unique values of \code{x$Decade}. Typically provided by calling function. See \code{stat_boxes_group}.}

\item{height}{character, stat box height. defaults to \code{"110px"}.}

\item{width.icon}{character, icon image width, defaults to \code{"90px"}.}

\item{text.size}{numeric, scaling for text size, defaults to \code{75}. See details.}

\item{value.size}{numeric, scaling for value size, defaults to \code{150}. See details.}

\item{output}{character, \code{"boxes"} (default) or \code{"list"}.
A list is useful for returning raw output, e.g. for sending to a rmarkdown report.}

\item{main_title}{character, typically contains html. Defaults to \code{"<h4>Aggregate period statistics</h4>"}.}

\item{clrs}{character, background colors of stat boxes. See details.}
}
\value{
a \code{shiny::tagList} containing a heading and a fluid row of six stat boxes.
}
\description{
Generate a set of six related stat boxes that summarize the distribution of values from a data frame.
}
\details{
\code{clrs} must be length-one (will be repeated six times), -two (alternating colors), or -six (all colors individually specified).
Colors must be provided in hex format, e.g., \code{#FFFFFF} or a valid Shiny Dashboard color.
This function is typically called by \code{stat_boxes_group} but may be called directly when groups of stat boxes are not needed.
\code{text.size} and \code{value.size} are length-six or length-one (repeated) vectors. Defaults are better suited to
\code{type="valueBox"} than \code{type="infoBox"}. Stat box icons are fixed images and come with the \code{apputils} package.
}
\examples{
#not run
}
