% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tolpal}
\alias{tolpal}
\title{Generate a vector of colors.}
\usage{
tolpal(n)
}
\arguments{
\item{n}{integer, number of colors requested in palette.}
}
\value{
a vector of colors.
}
\description{
Generate a vector of colors.
}
\examples{
#not run
}
