% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apputils.R
\name{update_toastr_css}
\alias{update_toastr_css}
\title{Update shinytoastr css}
\usage{
update_toastr_css(container = NULL, toast = NULL, position = "top-center")
}
\arguments{
\item{container}{list of style arguments for the container div. See details and example.}

\item{toast}{list of style arguments for the toast. See details and example.}

\item{position}{character, defaults to \code{"top-center"}.}
}
\value{
a character string of css.
}
\description{
Update toast css from shinytoastr package.
}
\details{
\code{apputils} already contains some toastr css overrides (loaded via \code{use_apputils}).
This function allows for injecting additonal or different css overrides for a specific toast container
that may not already be as specified by \code{apputils}. This is typically used to adjust the app intro toast,
hence the default for \code{position} is \code{"top-center"}.

Note that list names and values ay be quoted if necessary. See example.
Should be familiar with source toastr css in addition to running the example
in order to understand which elements apply to \code{container} vs. \code{toast}.
}
\examples{
cat(update_toastr_css(
  list('overflow-y'='auto', width='70\%', height='700px'),
  list(top='100px', margin='0 auto', left='115px')
))
}
