% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apputils.R
\name{use_apputils}
\alias{use_apputils}
\title{Initialize apputils}
\usage{
use_apputils()
}
\value{
HTML tags placed in the head of the app UI HTML file.
}
\description{
Call this function once near the top of the Shiny UI definition.
}
\details{
Calling \code{use_apputils} loads package css into the head of a shiny app.
}
\examples{
# See usage.
}
